check webgl for art only public view page
