+++ src/features/cosoul/art/CoSoulArtPublic.tsx
@@ -1,9 +1,13 @@
+import { useEffect, useRef, useState } from 'react';
+
 import { useQuery } from 'react-query';
 import { useParams } from 'react-router-dom';
 
 import { CosoulArtData } from '../../../../api/cosoul/art/[artTokenId]';
+import { Canvas, Flex } from 'ui';
 
 import { CoSoulArt } from './CoSoulArt';
+import { WebglMessage } from './WebglMessage';
 
 export const CoSoulArtPublic = ({ animate = true }: { animate?: boolean }) => {
   const params = useParams();
@@ -29,13 +33,44 @@ export const CoSoulArtPublic = ({ animate = true }: { animate?: boolean }) => {
     }
   );
 
+  const webglTest = useRef<HTMLCanvasElement>(null);
+  const [webglEnabled, setWebglEnabled] = useState(true);
+
+  useEffect(() => {
+    const canvas = webglTest.current;
+    const checkWebglEnabled = () => {
+      if (canvas) {
+        const webglEnabled = !!canvas.getContext('webgl2');
+        setWebglEnabled(webglEnabled);
+      }
+    };
+    checkWebglEnabled();
+  }, []);
+
   if (isLoading) {
-    return null;
+    return (
+      <Canvas
+        ref={webglTest}
+        css={{
+          position: 'absolute',
+          zIndex: -1,
+          left: -5000,
+        }}
+      />
+    );
   }
 
   if (!!data?.address && data?.pGive >= 0) {
     return (
-      <CoSoulArt pGive={data.pGive} address={data.address} animate={animate} />
+      <Flex css={{ position: 'relative' }}>
+        <WebglMessage webglEnabled={webglEnabled} />
+        <CoSoulArt
+          pGive={data.pGive}
+          address={data.address}
+          animate={animate}
+          webglEnabled={webglEnabled}
+        />
+      </Flex>
     );
   } else {
     return <>No CoSoul exists for this tokenId</>;
